Add files via upload
/* Momir Jovanovic
* HIRE DATE DOB TODAYS DATE ELIGIBILITY
* DATE: 05/31/2021
* SUBMIT CODE FOR CLAIMED SCENARIO
* ASSIGNMENT Due 05/23/2021 
* COP 2220C - 33940
* Professor: Jerry Reed
A company has very strict age limits for future employees. The company needs to find employees that will fit this age requirement
before or on the date of hire. The recruiter wants an easy way to calculate how old the applicant is, 
if they meet the required age, and how long until they do meet the required age. Create a program that:

1. Asks for a applicants DOB (month, day, and year)

2. Asks for a hiring date

3. Asks for an age requirement

4. Checks if the applicant fulfills the age requirement

5. If the applicant does not meet the age requirement, check to see how many days until they do

6. See if they will meet the age requirement by the date of hiring

7. Output whether the applicant is valid currently, valid by the time of hiring (include the amount of days until their birthday), or if they are not valid at all
*/