Create 5.10.1: Reading in a string too large for a C string 2220C.c
5.10.1: Reading in a string too large for a C string.
Run the program, which simply reads an input string and prints it one character at a time. Then, lengthen the input string beyond 10 characters, and run again. The program might work, if the extra memory locations being assigned don't matter. Try larger and larger strings, and see if the program fails (be sure to scroll to the bottom of the output to look for erroneous output or an error message).