Create 5.13.1 Array of strings example: Top 10 countries for most TV watched daily.c
5.13 Arrays and strings
Because C strings are stored using arrays of characters, an array of strings can be created using a two-dimensional char array. For example, char studentNames[5][30]; creates an array of 5 strings, in which each string contains up to 30 characters. Strings within the array are accessed using only the first array index, e.g., studentNames[0] will access the first string in the array. The following example illustrates the use of an array of strings.