Create 6.11.1: Modifying a string parameter: Spaces to hyphens.c
6.11.1: Modifying a string parameter: Spaces to hyphens.

Recall that a scanf() string argument did not require a prepended & as in scanf("%s", userName);, in contrast to integer or other argument types. The & is not needed because the compiler automatically passes an array as a pointer. The same is true for the userStr argument in fgets() above.

OUTPUT: 
Enter string with spaces: 
String with hyphens: Hello-there-everyone???