Create 7.4.1 A seat reservation system involving a struct, arrays, and functions.c
7.4 Structs, arrays, and functions: A seat reservation example
A programmer commonly uses structs, arrays, and functions together. Consider a program that allows a reservations agent to reserve seats for people, useful for a theater, an airplane, etc. The below program defines a Seat struct whose data members are a person's first name, last name, and the amount paid for the seat. The program declares an array of 5 seats to represent the theater, airplane, etc., initializes all seats to being empty (indicated by a first name of "empty"), and then allows a user to enter commands to print all seats, reserve a seat, or quit.

The programmer first defined several functions related to the Seat struct, such as checking if a seat is empty or printing a seat. The programmer then defined some functions related to an array of seat items. To distinguish, the programmer named the former starting with Seat and the latter starting with Seats.

The SeatMakeEmpty() function uses pass by pointer to update the information within an individual seat. Remember that to update a variable passed by pointer, the program must use the dereference operator * to access the value pointed to by the pointer. When the variable is a pointer to a structure, both the dereference operator and the member access operators must be used together. In this case, the member access operator has precedence, so parentheses are used to dereference the pointer first:

Construct 7.4.1: Dereferencing a pointer to a struct.
(*variableName).memberName

Feedback?
The programmer left a "FIXME" comment indicating that the program also requires the ability to delete a reservation. That functionality is straightforward to introduce, just requiring the user to enter a seat number and then making use of the existing SeatMakeEmpty() function.

Notice how main() is relatively clean, dealing mostly with the user commands, and then using functions to carry out the appropriate work. Actually, the "reserve seat" command could be improved; main() currently fills the reservation information (e.g., "Enter first name..."), but main() would be cleaner if it just called a function as SeatFillReservationInfo(&currSeat).

The seat reservation program loses all its information when exited. An improvement is to save all reservation information in a file. Commands 's' and 'g' would save and get information to/from a file, respectively.