Create 8.2.5 Checking to see if a pointer is null.c
someInt is located in memory at address 76. valPointer is assigned with NULL.
valPointer is passed to PrintValue(), so the valuePointer parameter is also NULL.
The if statement is true since valuePointer is NULL.
valPointer points to someInt, so calling PrintValue() assigns valuePointer with the address 76.
The if statement is false because valuePointer is no longer NULL. valuePointer points to the value 5, so 5 is output.

Null pointer
When a pointer is declared, the pointer variable holds an unknown address until the pointer is initialized. A programmer may wish to indicate that a pointer points to "nothing" by initializing a pointer to null. Null means "nothing". A pointer that is assigned with the macro NULL is said to be null. NULL is defined as 0 on most systems because 0 is not a valid memory address. Ex: int *maxValPointer = NULL; makes maxValPointer null.

In the animation below, the function PrintValue() only outputs the value pointed to by valuePointer if valuePointer is not null.Null pointer
When a pointer is declared, the pointer variable holds an unknown address until the pointer is initialized. A programmer may wish to indicate that a pointer points to "nothing" by initializing a pointer to null. Null means "nothing". A pointer that is assigned with the macro NULL is said to be null. NULL is defined as 0 on most systems because 0 is not a valid memory address. Ex: int *maxValPointer = NULL; makes maxValPointer null.

In the animation below, the function PrintValue() only outputs the value pointed to by valuePointer if valuePointer is not null.