Create 8.3.1 Malloc function.c
8.3 The malloc and free functions
//
Sometimes memory should be allocated while a program is running and should persist independently of any particular function. The malloc() function carries out such memory allocation.

malloc() is a function defined within the standard library, which can be used by adding #include <stdlib.h> to the beginning of a program. The malloc() function is named for memory allocation. malloc() allocates memory of a given number of bytes and returns a pointer (i.e., the address) to that allocated memory. A basic call to malloc() has the form:
Construct 8.3.1: Malloc function.
malloc(bytes)

malloc() takes a single argument specifying the number of bytes to allocate in memory. Thus, the programmer must determine the number of bytes needed to allocate space for the desired data type. But, as you may recall, the number of bytes for various data types (e.g., int) may vary across different computer systems. Fortunately, C provides a sizeof() operator that returns the number of bytes for a given data type. For example, on a system where an int is 32 bits, sizeof(int) returns 4, because 32-bits is 4 bytes. Calls to malloc() are typically combined with sizeof() using the form:

Construct 8.3.2: Malloc and sizeof functions.
malloc(sizeof(type))

malloc() returns a pointer to allocated memory using a void*, referred to as a void pointer. A void pointer is a special type of pointer that stores a memory address without referring to the type of variable stored at that memory location. The void pointer return type allows a single malloc() function to allocate memory for any data type. The pointer returned from malloc() can then be written into a particular pointer variable by casting the void pointer to the data type using the form:

Construct 8.3.3: Malloc return type.
pointerVariableName = (type*)malloc(sizeof(type));