Create 8.3.1 The malloc() operation.c
Construct 8.3.3: Malloc return type.
pointerVariableName = (type*)malloc(sizeof(type));

The following animation illustrates using malloc() to allocate memory for an int. int is used for introduction; malloc() is more commonly used to allocate memory for struct types, arrays, and strings.

Dereferencing the pointer variable allows the program to assign and access the value pointed to by the pointer.
Captions

Variable myPtr is a pointer of type int, which does not yet point to anything.
malloc(sizeof(int)) allocates memory for a single int. The void pointer returned by malloc is first cast to an int pointer before myPtr is assigned with that pointer.
Dereferencing the pointer variable allows the program to assign and access the value pointed to by the pointer.

The malloc() function returns NULL if the function failed to allocate memory. Such failure could happen if a program has used up all memory available to the program.