Create Figure 8.6.1 Dynamically allocating an array of integers.c
8.6 The malloc function for arrays and strings
Pointers are commonly used to allocate arrays just large enough to store a required number of data elements. Previously, the programmer had to declare arrays to have a fixed size, referred to as a statically allocated array. Statically allocated arrays may not use all the allocated memory due to the programmer not knowing the actual needed size before the program runs, thus creating larger-than-necessary arrays. Even then, the program might need a larger size.

Instead of statically allocating an array, malloc() can be used to allocate just enough memory for the array. Recall that arrays are stored in sequential memory locations. malloc() can be used to allocate a dynamically allocated array by determining the total number of bytes needed to store the desired number of elements, using the following form:

Construct 8.6.1: Dynamically allocated array.
pointerVariableName = (dataType*)malloc(numElements * sizeof(dataType))

Feedback?
When allocating an array, malloc() returns a pointer to memory location of the first element within the array. This memory location can be stored within a pointer variable declared as a pointer to the type of element within the array. For example, when dynamically allocating an array of integers, a pointer variable of integers "int*" is used. Notice then that an int* pointer can point to either a single integer or an array of multiple characters. A programmer must carefully keep track of how each pointer variable is utilized within the program.

The following program illustrates how to dynamically allocate an arrays of integers.