Create 8.6.2 Concatenating strings using a statically allocated array.c
The above program declares two statically allocated arrays named nameArr and greetingArr. Each array can store a string with 0 to 99 characters — keeping in mind the space needed to store the null character at the end of the string. However, if the user enters the name "Bob", the resulting string stored within the greeting array only requires 11 characters — 6 characters for "Hello ", 3 characters for the name "Bob", 1 character for the ".", and 1 character for the null character. Thus, 88 characters are unused in the greeting array. Likewise, the program fails if the user enters a very long name.
OUTPUT:
Enter name: Bill
Hello Bill.