Create 8.6.3 Concatenating strings using a dynamically allocated array.c
The nameArr array is used for reading the user input, so nameArr is statically allocated with a fixed size, defining a limit to the length of the name that the program can support.

In contrast, the size of the greeting is determined at runtime based upon the actual user-entered name length. char* greetingPtr; declares a char pointer variable named greetingPtr. Once the user has entered a name and the newline character at the end of the name string has been removed, strlen(nameArr) determines the number of characters within that name. In addition to the length of the string, 8 characters are needed for the greeting — 6 for the string "Hello ", 1 for the ".", and 1 for the end-of-string null character. Thus, the total number of bytes required for the greeting array can be determined using the expression strlen(nameArr) + 8) * sizeof(char). For example, if the user enters the name "Julia", the total number of characters allocated for the greeting array will be 5 + 8 = 13.

The program can then create the greeting array by first copying the string "Hello " to the greeting array using the statement strcpy(greetingPtr, "Hello ");. The statement strcat(greetingPtr, nameArr); then appends the user-entered name to the end of the string stored within greetingPtr. Finally, strcat(greetingPtr, "."); appends a "." to the end of the string.

To create a dynamically allocated copy of a string, malloc() can be used to create a character array with a size equal to the number of characters within the source string plus one character for the null character. The following program illustrates. As the length returned by strlen(nameArr) does not include the null character required at the end of a string, strlen(nameArr) + 1 is used to determine the number of characters required for the dynamically allocated string. A common error is to allocate only strlen chars for a copied string, forgetting the + 1 needed for the null character.

OUTPUT:
Enter name: Julia
Hello Julia.

...

Enter name: John Smith
Hello John Smith.