Create 8.6.4 Creating a dynamically allocated copy of a string.c
To create a dynamically allocated copy of a string, malloc() can be used to create a character array with a size equal to the number of characters within the source string plus one character for the null character. The following program illustrates. As the length returned by strlen(nameArr) does not include the null character required at the end of a string, strlen(nameArr) + 1 is used to determine the number of characters required for the dynamically allocated string. A common error is to allocate only strlen chars for a copied string, forgetting the + 1 needed for the null character.

OUTPUT:

Enter name: Julia
Hello Julia

...

Enter name: John Smith
Hello John Smith