Create 8.8.1 struct and function prototypes for vector ADT.c
A vector is an example of ADT that stores an ordered list of items (called elements) of a given data type, such as a vector of integers. Like an array, an individual element within the vector can be accessed using an index. However, unlike an array, a vector's size can be adjusted while a program is executing, an especially useful feature when the number of items that will be in the list is unknown at compile-time. To support such size adjustment, a vector ADT provides functions for common operations like creating the vector, inserting elements into the vector, removing elements from the vector, resizing the vector, and accessing elements at specific locations.

The following defines an ADT for a vector of integers — defining a new struct type named "vector" and the function prototypes for the vector's interface. These declarations are included in the header file for the ADT named "vector.h".

Vector ADT naming convention
For the vector ADT, this material uses a different convention for naming the functions. All functions supporting the vector ADT begin with "vector_", which is intended to indicate the functions are associated with the "vector" type. The following portion of the function corresponds to operation being performed on the vector ADT.

To use the vector, a program must first include the following: #include "vector.h". Then, a variable for a vector can be declared and used as shown in the below animation. The definition creates a vector variable named v. Data members within the vector struct are not initialized automatically. The vector_create() function is used to initialize the vector given the specified size of the vector. The function call vector_create(&v, 3) initializes the vector by allocating memory for three elements, each of type int, and initializes each of those elements to the value 0.

The function call vector_at(&v, 0) returns a pointer to the int element stored at location 0 of the vector v. A value can be written to this location by dereferencing the returned int pointer. For example, the statement *vector_at(&v, 0) = 119; assigns a value of 119 to the int element at location 0 of the vector v.

Notice that the first parameter for each of the vector's functions is a pointer to a vector ("vector*"). When calling each of these functions, this parameter will point to the specific vector on which the corresponding operation will be performed. Thus, a program can consist of multiple vectors, using the same set of the vector functions.