Create 8.11.1: Use of the four memory regions.c
8.11 Memory regions: Heap/Stack
A program's memory usage typically includes four different regions:

Code — The region where the program instructions are stored.
Static memory — The region where global variables (variables declared outside any function) as well as static local variables (variables declared inside functions starting with the keyword "static") are allocated. Static variables are allocated once and stay in the same memory location for the duration of a program's execution.
The stack — The region where a function's local variables are allocated during a function call. A function call adds local variables to the stack, and a return removes them, like adding and removing dishes from a pile; hence the term "stack." Because this memory is automatically allocated and deallocated, it is also called automatic memory.
The heap — The region where the malloc function allocates memory, and where the free function deallocates memory. The region is also called free store

The code regions store program instructions. myGlobal is a global variable and is stored in the static memory region. Code and static regions last for the entire program execution.
Function calls push local variables on the program stack. When main() is called, the variables myInt and myPtr are added on the stack.
malloc allocates memory on the heap for an int and returns the address of the allocated memory, which is assigned to myPtr. free deallocates memory from the heap.
Calling MyFct() grows the stack, pushing the function's local variables on the stack. Those local variables are removed from the stack when the function returns.
When main() completes, main's local variables are removed from the stack.