Create 9.1.1 Reading multiple values using a single scanf().c
9.1 Input parsing
This section describes features of the similar functions scanf, fscanf, and the soon-to-be-introduced sscanf, that support input parsing. The section illustrates using scanf, but the features apply to all three functions.

A programmer can control the way that input is read when using scanf(), a task known as input parsing. The format specifiers within the format string of scanf() can include format sub-specifiers. These sub-specifiers specify how the input will be read for that format specified. One of the most useful specifiers is the width specifier that can be used with the following form:

Construct 9.1.1: Format specifiers and sub-specifiers.
%(width)specifier

Feedback?
The width specifies the maximum number of character to read for the current format specifier. For example, the format string "%2d" will read in up to 2 characters -- in this case decimal digits -- converting the characters to the corresponding decimal value and storing that value into an integer variable.

A single scanf() statement can be used to read into multiple variables. The format string can include whitespace characters separating the format specifiers. These whitespace characters will cause the scanf() function to read all whitespace characters from the input until a non-whitespace character is reached. For example, the format string "%d %d" will read two decimal integers from the input separated by whitespace. That whitespace may be a single space, a newline, a space followed by a newline, or any combination thereof.

The following program uses a single scanf() statement to read two values for feet and inches, printing to equivalent distance in centimeters.