Create 9.1.2: An example of using non-whitespace characters in a format string to parse formatted input.c
User input often may include additional characters that are common to the format of the data being entered. For example, when receiving a time from a user, the programmer may prefer to allow users to use a common time format, such as "12:35 AM". In this example, the ':' is only used to format the data, separating the hour from the minute value.

The format string for scanf() can be configured to read the ':' character from the input but not store within a variable. scanf() will attempt to read any non-whitespace characters from the input. scanf() will only read the non-whitespace character if that character matches the provided user input.

Ex: the format string "%2d:%2d %2s" can be used to read in a time value:

The first format specifier "%2d" will read up to two decimal digits for the hour.
scanf() will then attempt to read a ':' character. If ':' is found in the user input, then ':' will be read and discarded.
The subsequent two format specifiers will read in the minutes and AM/PM setting.

OUTPUT:
Enter the time using the format: HH:MM AM/PM: 12:35 PM
In 12 hours it will be: 12:35 AM

...

Enter the time using the format: HH:MM AM/PM: 4:12AM
In 12 hours it will be: 04:12 PM

Importantly, as soon as scanf() is not able to match the format string, it will stop reading from the input. For example, if the user does not enter the ':' character, scanf() will immediately stop reading from the input. In such a situation the currMinutes and optAmPm variables will not be updated.