Create 9.1.3 Using the return value from scanf() to check for parsing errors.c
To check for such errors, the scanf() function returns an integer value for the number of items read using scanf() and stored within the specified variables. This return value can be checked to see if the user input matches the specified format. For example, if the user enters a valid time for the format string, scanf() will return 3. The following program extends the earlier example, printing an error message if the user input did not match the specified format string for all three format specifiers.