Create 9.1.4 Using sscanf() to parse a string.c
Sometimes a programmer wishes to read input data from a string rather than from the keyboard (standard input). The sscanf() function is used to read a sequence of characters from a C string, parsing the data stored within that string and storing the converted value within variables. The first argument to sscanf() is the string being read. The remaining arguments for sscanf() work the same way as the arguments for scanf(). Specifically, the second argument for the sscanf() function is the format string that specifies the type of value to be read using a format specifier. The argument following the format string is the location to store the values that are read.

Unlike the scanf() function that continues reading from the user input where the previous scanf() stopped, sscanf() always starts at the beginning of the specified string. In addition, the contents of the string being read are not modified by sscanf(). The following program illustrates.

A common use of scanf() is to process user input line-by-line. The following program reads in the line as a string, and then extracts individual data items from that string.