Create 9.1.5 Using a sscanf() to parse a line of input text.c
A common use of scanf() is to process user input line-by-line. The following program reads in the line as a string, and then extracts individual data items from that string.

The program uses fgets() to read an input line into a string. Recall that C strings are implemented using character arrays. As the size of the character array -- or string -- must be known before calling fgets(), if the user enters a line of text that is longer than the length of that string, care must be taken to ensure the user input is not written to an out of bounds index.

The second argument to the fgets() function is an integer value specifying the maximum number of characters to write to the specified string. Using this input correctly ensures fgets() will not write to out of range values for the specified string. For example, if inputBuffer is declared as char inputBuffer[100], the statement fgets(inputBuffer, 100, stdin); will ensure that no more than 100 characters are written to the string inputBuffer. Additionally, fgets() will ensure that the null character will be written to the end of the string read.

Similarly, when parsing a string -- or user input -- to read a string, the width sub-specifier of the "%s" format specifier should be used. Recall that the width sub-specifier specifies the maximum number of characters to read. If myString is defined char myString[50], the format specifier "%49s" can be used to ensure no more than 49 characters are read from the input, leaving one space for the null character at the end of the string.

A good practice is to always use the width sub-specifier when reading strings using scanf(), fscanf(), or sscanf().

OUTPUT:

Enter "firstname lastname age" on each line
("Exit" as firstname exits).

Amy Smith 19
   First name: Amy
   Last  name: Smith
   Age: 19

Mike Smith 24
   First name: Mike
   Last  name: Smith
   Age: 24

No Age
Invalid entry. Please try again.

Exit
Exiting.