Create 9.2 Output formatting.c
9.2 Output formatting
A programmer can adjust the way that a program's output appears, a task known as output formatting. The format specifiers within the format string of printf() and fprintf() can include format sub-specifiers. These sub-specifiers specify how a value stored within a variable will be printed in place of a format specifier.

Floating-point values
Formatting floating-point output is commonly done using sub-specifiers. A sub-specifier provides formatting options for a format specifier and are included between the % and format specifier character. Ex: The .1 sub-specifier in printf("%.1f", myFloat); causes the floating-point variable myFloat to be output with only 1 digit after the decimal point; if myFloat is 12.34, the output would be 12.3. Format specifiers and sub-specifiers use the following form:

Construct 9.2.1: Format specifiers and sub-specifiers.
%(flags)(width)(.precision)specifier

Feedback?
The table below summarizes the floating-point sub-specifiers available. Assume myFloat has a value of 12.34. Recall that "%f" is used for float values, "%lf" is used for double values, "%e" is used to display float values in scientific notation, and "%le" is used to display double values in scientific notation.