Create 9.5.1 Input from a file.c
9.5 File input and output
Sometimes a program should get input from a file rather than from a user typing on a keyboard. To achieve this, a programmer can open another input file, rather than the predefined input file stdin that comes from the standard input (keyboard). That new input file can then be used with fscanf() just like using scanf() with the stdin file, as the following program illustrates. Assume a text file exists named myfile.txt with the contents shown (created for example using Notepad on a Windows computer or using TextEdit on a Mac computer).


Six lines are needed for input from a file, highlighted above.

The #include <stdio.h> enables use of FILE* variables and supporting functions.
A new FILE* variable has been declared: FILE* inputFile;.FilePointer
The line inputFile = fopen("myfile.txt", "r"); then opens the file for reading and associates the file with the FILE*. The first argument to fopen() is a string with the name of the file to open. The second argument of fopen() is a string indicating the file mode, which specifies if the file should be open for reading or writing. The string "r" indicates the file should be open for reading, referred to as read mode. Upon success, fopen() will return a pointer to the FILE structure for the file that was opened. If fopen() could not open the file, it will return NULL.
Because of the high likelihood that the open fails, usually because the file does not exist or is in use by another program, the program checks whether the open was successful using if (inputFile == NULL).
The successfully opened input file is read from using fscanf(), e.g., using fscanf(inFile, "%d %d", &num1, &num2); to read two integers into num1 and num2.
Finally, when done using the file, the program closes the file using fclose(inputFile);.
A common error is to specify the file mode as a character (e.g. 'r') rather than a string (e.g."r"). Another common error is a mismatch between the variable data type and the file data, e.g., if the data type is int but the file data is "Hello".