Create 9.5.2 Program that reads data from myfile.txt into an array.c
The following provides another example wherein the program reads items into a dynamically allocated array. For this program, myfile.txt's first entry must be the number of numbers to read, followed by those numbers, e.g., 5 10 20 40 80 1.
A program can read varying amounts of data in a file by using a loop that reads until the end of the file has been reached, as follows.

The feof() function returns 1 if the previous read operation reached the end of the file. Errors may be encountered while attempting to read from a file, including end-of-file, corrupt data, etc. So, a program should check that each read was successful before using the variable to which the data read was assigned. fscanf() returns the number of items read from the file and assigned to a variable, which can be checked to determine if the read operation was successful. Ex: if( fscanf(inFile, "%d", &fileNum) == 1 ) {...} checks that fscanf() read and assigned a value to fileNum.