Create 9.5.3 Reading a varying amount of data from a file.c
A program can read varying amounts of data in a file by using a loop that reads until the end of the file has been reached, as follows.

The feof() function returns 1 if the previous read operation reached the end of the file. Errors may be encountered while attempting to read from a file, including end-of-file, corrupt data, etc. So, a program should check that each read was successful before using the variable to which the data read was assigned. fscanf() returns the number of items read from the file and assigned to a variable, which can be checked to determine if the read operation was successful. Ex: if( fscanf(inFile, "%d", &fileNum) == 1 ) {...} checks that fscanf() read and assigned a value to fileNum.

OUTPUT:
myfile.txt with variable number of integers:
111
222
333
444
555
Opening file myfile.txt.
Reading and printing numbers.
num: 111
num: 222
num: 333
num: 444
num: 555
Closing file myfile.txt.
Similarly, a program may write output to a file rather than to standard output, as shown below. To open an output file, the string "w" is used as the file mode within the call to fopen(), referred to as write mode. Using the write mode, if a file with specified name already exists, that file will be replaced with the newly created file.