Create 9.5.4 Sample code for writing to a file.c
Similarly, a program may write output to a file rather than to standard output, as shown below. To open an output file, the string "w" is used as the file mode within the call to fopen(), referred to as write mode. Using the write mode, if a file with specified name already exists, that file will be replaced with the newly created file.

OUTPUT:
Contents of myoutfile.txt after running the program:
Hello
1 2 3