Create 9.6.1 Sample code for writing an array of integers one-by-one to a binary file.c
9.6 Binary file I/O
Programs can also access files using a binary file mode. Using the binary file mode, a program can directly copy data in the program's memory to and from the file. This mode is referred to as binary mode because reading and writing to the file will copy the contents to and from memory bit by bit.

Binary files provide several advantages.

Reading from and writing to binary files is very efficient because these operations directly copy data without first converting the data into a human readable format. For example, consider writing a 32-bit integer value, such as -13645766, to an output file. Using a character file, this 32-bit integer would first need to be converted to the sequence of characters "-13645766" representing that integer value. When writing to a binary file, this conversion is not needed.
Binary files can be more space efficient than character files. Again, consider writing the same integer value -13645766 to an output file. Using a binary file, the value requires 32-bits -- or 4 bytes -- as the integer value is stored in memory using 4 bytes. Using a character file would require 9 bytes -- one byte for each character within "-13645766".
Binary files also have disadvantages. The main disadvantage of binary files is that they cannot be easily read or edited by humans. A good practice is to only use binary files when those files will never be directly read or edited by a user.

To open a file using a binary mode, the file mode string in the call to fopen() should include the character b. To open a file for reading using the binary mode, the file mode string "rb" would be used. Similarly, to open a file for writing using the binary mode, the file mode string "wb" would be used.

A program can write to a binary file using the fwrite() function using the following form:

Construct 9.6.1: fwrite() function.
fwrite(variablePointer, sizeof(type), numElements, outputFile)

Feedback?
The first argument of fwrite() is a pointer to the variable, or memory location, that will be copied and written to the binary file. The second and third arguments specify the size in bytes of each element and the number of elements, respectively, being written to the file. The sizeof() operation can be used to determine the number of bytes required for the data type of the variable or array being written. The final argument to fwrite() is a FILE* for the output file being written. fwrite() will return the total number of bytes written to the output file.

The following demonstrates the use of fwrite() to write several user-entered integer numbers to a binary file.
Try 9.6.1: Viewing a binary file in a text editor.
Run the above program and try opening the resulting myfile.bin file using a text editor. Notice that the contents of the file are not easily readable.